+++ php/src/Application/Command/Reindex.php
@@ -80,6 +80,7 @@ class Reindex extends AbstractCommand
     {
         $optionCollection->add('source+', 'The file or directory to index. Can be passed multiple times to process multiple items at once.')->isa('string');
         $optionCollection->add('exclude+', 'An absolute path to exclude. Can be passed multiple times.')->isa('string');
+        $optionCollection->add('extension+', 'An extension (without leading dot) to index. Can be passed multiple times.')->isa('string');
         $optionCollection->add('stdin?', 'If set, file contents will not be read from disk but the contents from STDIN will be used instead.');
         $optionCollection->add('v|verbose?', 'If set, verbose output will be displayed.');
         $optionCollection->add('s|stream-progress?', 'If set, progress will be streamed. Incompatible with verbose mode.');
@@ -99,7 +100,8 @@ class Reindex extends AbstractCommand
             isset($arguments['stdin']),
             isset($arguments['verbose']),
             isset($arguments['stream-progress']),
-            isset($arguments['exclude']->value) ? $arguments['exclude']->value : []
+            isset($arguments['exclude']->value) ? $arguments['exclude']->value : [],
+            isset($arguments['extension']->value) ? $arguments['extension']->value : []
         );
 
         return $this->outputJson($success, []);
@@ -111,11 +113,18 @@ class Reindex extends AbstractCommand
      * @param bool     $showOutput
      * @param bool     $doStreamProgress
      * @param string[] $excludedPaths
+     * @param string[] $extensionsToIndex
      *
      * @return bool
      */
-    public function reindex(array $paths, $useStdin, $showOutput, $doStreamProgress, array $excludedPaths = [])
-    {
+    public function reindex(
+        array $paths,
+        $useStdin,
+        $showOutput,
+        $doStreamProgress,
+        array $excludedPaths = [],
+        array $extensionsToIndex = ['php']
+    ) {
         if ($useStdin) {
             if (count($paths) > 1) {
                 throw new UnexpectedValueException('Reading from STDIN is only possible when a single path is specified!');
@@ -143,7 +152,7 @@ class Reindex extends AbstractCommand
                     $sourceOverrideMap[$paths[0]] = $this->getSourceCodeHelper()->getSourceCode($paths[0], true);
                 }
 
-                $this->getProjectIndexer()->index($paths, $excludedPaths, $sourceOverrideMap);
+                $this->getProjectIndexer()->index($paths, $extensionsToIndex, $excludedPaths, $sourceOverrideMap);
             } catch (Indexing\IndexingFailedException $e) {
                 $success = false;
             }